Move ActionEvent to separate file
