+++ WordPress/src/main/java/org/wordpress/android/ui/jpfullplugininstall/install/ActionEvent.kt
@@ -1,4 +1,13 @@
 package org.wordpress.android.ui.jpfullplugininstall.install
 
+import org.wordpress.android.fluxc.model.SiteModel
+import org.wordpress.android.ui.accounts.HelpActivity
+
 sealed class ActionEvent {
+    data class ContactSupport(
+        val origin: HelpActivity.Origin,
+        val selectedSite: SiteModel?,
+    ) : ActionEvent()
+
+    object Dismiss : ActionEvent()
 }
